Fixes bugs and expands API of HashTable
This iteration of this container is the result of me testing it with my programming language project. There are no longer double-free()s when Hashtables are moved or copied, and the API is much larger and more capable.

I'm sure I'll have to tinker at it more, but this is a really good important so far :-)